+++ Tests/Filters/RegexReplaceTest.php
@@ -6,6 +6,7 @@
   use Fiv\Form\Filter\CallbackFilter;
   use Fiv\Form\Filter\RegexReplace;
   use Fiv\Form\Form;
+  use Fiv\Form\FormData;
 
   /**
    * @package Tests\Form\Filters
@@ -29,10 +30,10 @@
 
       $form->input('text')->addFilter(new RegexReplace('!\s{2,}!', ' '))->addFilter(new CallbackFilter('trim'));
 
-      $form->setData([
+      $form->handle(new FormData('post', [
         'test_form' => 1,
         'text' => 'hello    world         ',
-      ]);
+      ]));
 
       $this->assertTrue($form->isValid());