+++ Tests/Validator/RequiredTest.php
@@ -3,6 +3,7 @@
   namespace Tests\Fiv\Form\Validator;
 
   use Fiv\Form\Form;
+  use Fiv\Form\FormData;
   use Fiv\Form\Validator\Required;
   use PHPUnit_Framework_TestCase;
 
@@ -20,28 +21,26 @@
       $form->input('login')
         ->addValidator($validator);
 
-      $form->setData([
+      $form->handle(new FormData('post', [
         $form->getUid() => 1,
         'login' => 'testLogin',
-      ]);
+      ]));
 
       $this->assertTrue($form->isValid());
       $this->assertEmpty($validator->getErrors());
 
-
-      $form->setData([
+      $form->handle(new FormData('post', [
         $form->getUid() => 1,
         'login' => '',
-      ]);
+      ]));
 
       $this->assertFalse($form->isValid());
       $this->assertEquals('Test error message', $validator->getFirstError());
 
-
-      $form->setData([
+      $form->handle(new FormData('post', [
         $form->getUid() => 1,
         'login' => '0',
-      ]);
+      ]));
       $this->assertTrue($form->isValid());
     }