+++ src/Element/CheckboxList.php
@@ -2,6 +2,9 @@
 
   namespace Fiv\Form\Element;
 
+  use Fiv\Form\FormData;
+
+
   /**
    * Class TextArea
    * Generate
@@ -14,7 +17,16 @@
    * @author  Ivan Shcherbak <dev@funivan.com>
    * @package Fiv\Form
    */
-  class CheckboxList extends \Fiv\Form\Element\Multiple {
+  class CheckboxList extends Multiple {
+
+    /**
+     * @inheritdoc
+     */
+    public function handle(FormData $data) {
+      $values = (array) $data->get($this->getName(), []);
+      $this->setValue($values);
+    }
+
 
     /**
      * @param array|string $data
@@ -34,6 +46,23 @@
 
       $this->value = array_values($data);
       return $this;
+    }
+
+
+    /**
+     * @return array
+     */
+    public function getValue() {
+      return (array) parent::getValue();
+    }
+
+
+    /**
+     * @param string $optionKey
+     * @return bool
+     */
+    public function isChecked($optionKey) {
+      return in_array($optionKey, $this->getValue());
     }